requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
distlib==0.3.9
filelock==3.17.0
idna==3.10
platformdirs==4.3.6
python-dotenv==1.0.0
requests==2.32.3
urllib3==2.3.0
virtualenv==20.29.2
